+++ native/rust/crates/ripdpi-monitor-engine/src/execution/runtime.rs
@@ -2,6 +2,7 @@ mod adaptive_freeze;
 mod contracts;
 mod launch;
 mod preparation;
+mod tcp_launch;
 mod timeout;
 mod unavailable;
 mod warmup;
@@ -12,7 +13,7 @@ use std::time::Duration;
 pub(crate) use adaptive_freeze::freeze_adaptive_fake_ttl_for_probe;
 pub use contracts::{CandidateProbeRuntime, CandidateRuntimeError, CandidateRuntimeLauncher, PreparedCandidateRuntime};
 pub use launch::probe_runtime_transport;
-pub(crate) use launch::probe_tcp_runtime_transport;
+pub(crate) use tcp_launch::probe_tcp_runtime_transport;
 pub use unavailable::UnavailableCandidateRuntimeLauncher;
 pub use warmup::run_candidate_warmup;
