+++ native/rust/crates/ripdpi-monitor-engine/src/execution/runtime/launch.rs
@@ -23,17 +23,3 @@ pub fn probe_runtime_transport(
         }
     }
 }
-
-pub(crate) fn probe_tcp_runtime_transport(
-    launcher: &dyn CandidateRuntimeLauncher,
-    spec: &StrategyCandidateSpec,
-    runtime_context: Option<&ProxyRuntimeContext>,
-) -> Result<Box<dyn CandidateProbeRuntime>, CandidateRuntimeError> {
-    let mut probe_spec = spec.clone();
-    // TCP candidate probes never exercise SOCKS5 UDP ASSOCIATE. Keeping the
-    // user's default here makes a TCP-only relay fail validation before any
-    // target is tested. Override only the ephemeral runtime spec so the
-    // candidate's persisted/recommended configuration remains unchanged.
-    probe_spec.config.protocols.udp_associate_enabled = Some(false);
-    probe_runtime_transport(launcher, &probe_spec, runtime_context)
-}
